chore: Factor out uuid helpers
