+++ src/queries/uuid.ts
@@ -1,18 +1,6 @@
 import { queryOptions } from "@tanstack/react-query";
 import { env } from "#env.ts";
-
-export const normalizeUUID = (uuid: string) => {
-    const cleaned = uuid.replace(/[^\da-fA-F]/g, "").toLowerCase();
-
-    if (cleaned.length !== 32) {
-        throw new Error(`Invalid UUID: ${uuid}`);
-    }
-
-    return `${cleaned.slice(0, 8)}-${cleaned.slice(8, 12)}-${cleaned.slice(
-        12,
-        16,
-    )}-${cleaned.slice(16, 20)}-${cleaned.slice(20)}`;
-};
+import { normalizeUUID } from "#helpers/uuid.ts";
 
 export const getUUIDQueryOptions = (username: string) =>
     queryOptions({
